backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
crewai==0.30.11
langchain==0.1.20
google-generativeai==0.5.4
httpx==0.27.0
